+++ anki-web-browser/src/__init__.py
@@ -15,6 +15,7 @@ def logToConsole(*args, **kargs):
 try:
     # if not '--test' in sys.argv:
 
+    # Uncomment to produce more logs on console
     from .core import Feedback
     Feedback.log = logToConsole